+++ automated-tests/src/dali/utc-Dali-ImageActor.cpp
@@ -864,7 +864,7 @@ int UtcDaliImageActorPropertyIndices(void)
   Property::IndexContainer indices;
   imageActor.GetPropertyIndices( indices );
   DALI_TEST_CHECK( indices.Size() > basicActor.GetPropertyCount() );
-  DALI_TEST_EQUALS( indices.Size(), imageActor.GetPropertyCount(), TEST_LOCATION );
+  DALI_TEST_EQUALS( (unsigned int)indices.Size(), imageActor.GetPropertyCount(), TEST_LOCATION );
   END_TEST;
 }
