+++ automated-tests/src/dali/utc-Dali-ShaderEffect.cpp
@@ -751,7 +751,7 @@ int UtcDaliShaderEffectPropertyIndices(void)
   Property::IndexContainer indices;
   effect.GetPropertyIndices( indices );
   DALI_TEST_CHECK( indices.Size() );
-  DALI_TEST_EQUALS( indices.Size(), effect.GetPropertyCount(), TEST_LOCATION );
+  DALI_TEST_EQUALS( (unsigned int)indices.Size(), effect.GetPropertyCount(), TEST_LOCATION );
   END_TEST;
 }
